+++ spec/kettle/dev/release_cli_spec.rb
@@ -42,29 +42,29 @@ RSpec.describe Kettle::Dev::ReleaseCLI do
       end
     end
 
-    describe "#run_cmd! (signing env injection)" do
+    describe "#run_cmd! (signing env injection)", :real_release_rake do
       it "prefixes SKIP_GEM_SIGNING for 'bundle exec rake build' when env set" do
         stub_env("SKIP_GEM_SIGNING" => "true")
-        expect(cli).to receive(:system).with(kind_of(Hash), "SKIP_GEM_SIGNING=true bundle exec rake build").and_return(true)
+        expect(described_class).to receive(:system).with(kind_of(Hash), "SKIP_GEM_SIGNING=true bundle exec rake build").and_return(true)
         cli.send(:run_cmd!, "bundle exec rake build")
       end
 
       it "prefixes SKIP_GEM_SIGNING for 'bundle exec rake release' when env set" do
         stub_env("SKIP_GEM_SIGNING" => "true")
-        expect(cli).to receive(:system).with(kind_of(Hash), "SKIP_GEM_SIGNING=true bundle exec rake release").and_return(true)
+        expect(described_class).to receive(:system).with(kind_of(Hash), "SKIP_GEM_SIGNING=true bundle exec rake release").and_return(true)
         cli.send(:run_cmd!, "bundle exec rake release")
       end
 
       it "does not prefix when SKIP_GEM_SIGNING is not set" do
         # ensure var is not present
         stub_env("SKIP_GEM_SIGNING" => nil)
-        expect(cli).to receive(:system).with(kind_of(Hash), "bundle exec rake build").and_return(true)
+        expect(described_class).to receive(:system).with(kind_of(Hash), "bundle exec rake build").and_return(true)
         cli.send(:run_cmd!, "bundle exec rake build")
       end
 
       it "does not prefix unrelated commands" do
         stub_env("SKIP_GEM_SIGNING" => "true")
-        expect(cli).to receive(:system).with(kind_of(Hash), "bin/rake").and_return(true)
+        expect(described_class).to receive(:system).with(kind_of(Hash), "bin/rake").and_return(true)
         cli.send(:run_cmd!, "bin/rake")
       end
     end